src/Dapr-azure-queue/requirements.txt
dapr
requests
dapr.ext.grpc
flask
Pillow
azure.storage.blob